move item-header component at its right place in the project structure
