+++ src/app/modules/item/components/access-code-view/access-code-view.component.scss
